ForbiddenNames: remove sniffing for `class ... extends $forbidden`
The `ForbiddenNames` sniff is intended to sniff for use of reserved keywords when used to declare namespaces/classes/functions/etc.

Sniffing for `class ... extends $forbidden` does not identify the root cause, but only points out that _somewhere else_ there must be a class which is named after a reserved keyword.

In other words, sniffing for `extends` does not actually help developers pinpoint the real problem.

That is quite apart from the fact that `T_EXTENDS` should be handled differently if it would continue to be sniffed for, as the extended classname may be composed of a namespace + classname (and namespace separators) and non of these parts is allowed to use a reserved keyword. Currently the sniff only looks at the very first part it encounters, which may be a namespace òr a classname and may well hide reserved keywords, i.e. `MyName\catch\MyClass`.