WAIT - PHPCSUtils alpha 4 | :sparkles: PHP 8.0 | New `NewNamespacedNameSpacingRestrictions` sniff
Detect code which doesn't comply with the new namespaced name spacing rules in PHP 8.0.

> As of PHP 8.0, whitespace [red: and comments] is not permitted between namespace separators.
> If it occurs, the namespace separator will be parsed as `T_NS_SEPARATOR`, which will subsequently
> lead to a parse error.
> It is not possible to allow whitespace, because namespaced names commonly occur next to keywords.

The above also implies (and the RFC/PHP PR confirms) that a space or comment between a keyword and a namespaced name after it, is now a requirement.

This new sniff will find violations against these new spacing rules and report them.

The sniff is quite complicated, especially as keywords used in namespaced names are allowed as of PHP 8.0, but these keywords will still be tokenized as keywords in PHP < 8. I've tried to be comprehensive about handling this, but expect potential false positives/negatives when keywords are used in namespaced names.

Includes extensive unit tests.

Ref:
* https://wiki.php.net/rfc/namespaced_names_as_token