BaseSniffTestCase: add `Before/After[Class]` attributes
PHPUnit 10 introduced attributes as replacements for docblock annotations.
PHPUnit 11 deprecates the use of docblock annotations in favour of attributes.

If both an attribute as well as an annotation are found, no PHPUnit deprecation warning will be thrown.

While not strictly necessary yet, this commit adds the `Before/After*` attributes in the `BaseSniffTestCase` class to reduce the amount of deprecation notices somewhat (Before: > 1050, after: ~694).

_The rest of the "annotations to attributes" conversion for the test suite is left for the update to PHPUnit 12 (if and when)._

Note: The `@before/after*` annotations remain as the tests also still need to run on PHP 5.4 - 8.0 using PHPUnit 4.x - 9.x.
These can be removed once the codebase has a PHP 8.1/PHPUnit 10 minimum requirement.
Alternatively, once support for PHP < 7.2 / PHPUnit < 8.0 is dropped, the fixture methods can be reverted to use the canonical `setUp/tearDown[BeforeClass|AfterClass]` names again.

Note: due to the syntax for attributes, these can be safely added as they are ignored as comments on PHP < 8.0.
Along the same line, if there is no "listener" for the attributes (PHP 8.0/PHPUnit 9.x), they are ignored by PHP as well.