PHP 8.4 | :sparkles: New `PHPCompatibility.Classes.NewFinalProperties` sniff
As of PHP 8.4, properties can be declared as `final`.

Note: the `final` keyword is not (yet) supported for constructor promoted properties. Support for this will be added in PHP 8.5 (and the sniff will need to be updated once PHPCS supports this).

This commit adds a new sniff to detect final properties as supported in PHP 8.4.

Includes tests.
Includes documentation.

Refs:
* https://wiki.php.net/rfc/property-hooks#final_hooks
* php/php-src 13455
* https://github.com/php/php-src/commit/780a8280d23453ebab5df0dbc35ac897c07c1f0d