AbstractInitialValueSniff: start using PHPCSUtils 1.1.0 getDeclaredProperties()
PHPCSUtils 1.1.0 introduces a new `ObjectDeclarations::getDeclaredProperties()` method which can retrieve a list of all properties declared in an OO structure.

This method is highly optimized and will cache its results, which means that if multiple sniffs use the method, the results will be instantaneously returned on subsequent uses.

Includes an additional test via one of the sniffs using the abstract.