GH Actions: add workflow to label PRs which are in a conflict state
This commit adds a new workflow which automatically checks for merge conflicts in open PRs.

This allows to see from the PR overview page which PRs need rebase/conflict resolution and are not currently in a mergable state.

The selected action runner also has an option to post a comment when a conflict is detected as well as when the conflict has been resolved. While not currently used, this feature might be useful in the future.

Refs:
* https://github.com/eps1lon/actions-label-merge-conflict